Refactored the package and update the theming to support custom colors
* FFTheme is refactored to support Libraries
