+++ lib/flutterflow_ui.dart
@@ -1,4 +1,6 @@
 library flutterflow_ui;
 
-export 'src/flutter_flow/flutter_flow.dart';
-export 'src/flutter_flow/random_data_util.dart';
+export 'src/themes/flutter_flow_theme.dart';
+export 'src/utils/flutter_flow_utils.dart';
+export 'src/utils/random_data.dart';
+export 'src/widgets/flutter_flow_widgets.dart';
