+++ lib/src/themes/custom_color.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class CustomColor {
+  final String name;
+  final Color color;
+
+  const CustomColor({
+    required this.name,
+    required this.color,
+  });
+}
