+++ lib/src/widgets/flutter_flow_choice_chips.dart
@@ -1,4 +1,4 @@
-import 'form_field_controller.dart';
+import '../utils/form_field_controller.dart';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
