+++ lib/src/widgets/flutter_flow_drop_down.dart
@@ -1,7 +1,6 @@
 import 'package:dropdown_button2/dropdown_button2.dart';
 
-
-import 'form_field_controller.dart';
+import '../utils/form_field_controller.dart';
 import 'package:flutter/material.dart';
 
 class FlutterFlowDropDown<T> extends StatefulWidget {
@@ -38,16 +37,16 @@ class FlutterFlowDropDown<T> extends StatefulWidget {
     this.labelText,
     this.labelTextStyle,
   }) : assert(
-  isMultiSelect
-      ? (controller == null &&
-      onChanged == null &&
-      multiSelectController != null &&
-      onMultiSelectChanged != null)
-      : (controller != null &&
-      onChanged != null &&
-      multiSelectController == null &&
-      onMultiSelectChanged == null),
-  );
+          isMultiSelect
+              ? (controller == null &&
+                  onChanged == null &&
+                  multiSelectController != null &&
+                  onMultiSelectChanged != null)
+              : (controller != null &&
+                  onChanged != null &&
+                  multiSelectController == null &&
+                  onMultiSelectChanged == null),
+        );
 
   final FormFieldController<T?>? controller;
   final FormFieldController<List<T>?>? multiSelectController;
@@ -107,21 +106,21 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
   }
 
   Map<T, String> get optionLabels => Map.fromEntries(
-    widget.options.asMap().entries.map(
-          (option) => MapEntry(
-        option.value,
-        widget.optionLabels == null ||
-            widget.optionLabels!.length < option.key + 1
-            ? option.value.toString()
-            : widget.optionLabels![option.key],
-      ),
-    ),
-  );
+        widget.options.asMap().entries.map(
+              (option) => MapEntry(
+                option.value,
+                widget.optionLabels == null ||
+                        widget.optionLabels!.length < option.key + 1
+                    ? option.value.toString()
+                    : widget.optionLabels![option.key],
+              ),
+            ),
+      );
 
   EdgeInsetsGeometry get horizontalMargin => widget.margin.clamp(
-    EdgeInsetsDirectional.zero,
-    const EdgeInsetsDirectional.symmetric(horizontal: double.infinity),
-  );
+        EdgeInsetsDirectional.zero,
+        const EdgeInsetsDirectional.symmetric(horizontal: double.infinity),
+      );
 
   late void Function() _listener;
   final TextEditingController _textEditingController = TextEditingController();
@@ -176,9 +175,9 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
 
   bool _useDropdown2() =>
       widget.isMultiSelect ||
-          widget.isSearchable ||
-          !widget.isOverButton ||
-          widget.maxHeight != null;
+      widget.isSearchable ||
+      !widget.isOverButton ||
+      widget.maxHeight != null;
 
   Widget _buildDropdownWidget() =>
       _useDropdown2() ? _buildDropdown() : _buildLegacyDropdown();
@@ -195,7 +194,9 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
       dropdownColor: widget.fillColor,
       focusColor: Colors.transparent,
       decoration: InputDecoration(
-        labelText: widget.labelText == null || widget.labelText!.isEmpty ? null : widget.labelText,
+        labelText: widget.labelText == null || widget.labelText!.isEmpty
+            ? null
+            : widget.labelText,
         labelStyle: widget.labelTextStyle,
         border: widget.hidesUnderline
             ? InputBorder.none
@@ -211,68 +212,66 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
   List<DropdownMenuItem<T>> _createMenuItems() => widget.options
       .map(
         (option) => DropdownMenuItem<T>(
-      value: option,
-      child: Padding(
-        padding: _useDropdown2() ? horizontalMargin : EdgeInsets.zero,
-        child: Text(optionLabels[option] ?? '', style: widget.textStyle),
-      ),
-
-    ),
-  )
+          value: option,
+          child: Padding(
+            padding: _useDropdown2() ? horizontalMargin : EdgeInsets.zero,
+            child: Text(optionLabels[option] ?? '', style: widget.textStyle),
+          ),
+        ),
+      )
       .toList();
 
   List<DropdownMenuItem<T>> _createMultiselectMenuItems() => widget.options
       .map(
         (item) => DropdownMenuItem<T>(
-      value: item,
-      // Disable default onTap to avoid closing menu when selecting an item
-      enabled: false,
-      child: StatefulBuilder(
-        builder: (context, menuSetState) {
-          final isSelected =
-              multiSelectController.value?.contains(item) ?? false;
-          return InkWell(
-            onTap: () {
-              multiSelectController.value ??= [];
-              isSelected
-                  ? multiSelectController.value!.remove(item)
-                  : multiSelectController.value!.add(item);
-              multiSelectController.update();
-              // This rebuilds the StatefulWidget to update the button's text.
-              setState(() {});
-              // This rebuilds the dropdownMenu Widget to update the check mark.
-              menuSetState(() {});
-            },
-            child: Container(
-              height: double.infinity,
-              padding: horizontalMargin,
-              child: Row(
-                children: [
-                  if (isSelected)
-                    const Icon(Icons.check_box_outlined)
-                  else
-                    const Icon(Icons.check_box_outline_blank),
-                  const SizedBox(width: 16),
-                  Expanded(
-                    child: Text(
-                      optionLabels[item]!,
-                      style: widget.textStyle,
-                    ),
+          value: item,
+          // Disable default onTap to avoid closing menu when selecting an item
+          enabled: false,
+          child: StatefulBuilder(
+            builder: (context, menuSetState) {
+              final isSelected =
+                  multiSelectController.value?.contains(item) ?? false;
+              return InkWell(
+                onTap: () {
+                  multiSelectController.value ??= [];
+                  isSelected
+                      ? multiSelectController.value!.remove(item)
+                      : multiSelectController.value!.add(item);
+                  multiSelectController.update();
+                  // This rebuilds the StatefulWidget to update the button's text.
+                  setState(() {});
+                  // This rebuilds the dropdownMenu Widget to update the check mark.
+                  menuSetState(() {});
+                },
+                child: Container(
+                  height: double.infinity,
+                  padding: horizontalMargin,
+                  child: Row(
+                    children: [
+                      if (isSelected)
+                        const Icon(Icons.check_box_outlined)
+                      else
+                        const Icon(Icons.check_box_outline_blank),
+                      const SizedBox(width: 16),
+                      Expanded(
+                        child: Text(
+                          optionLabels[item]!,
+                          style: widget.textStyle,
+                        ),
+                      ),
+                    ],
                   ),
-                ],
-              ),
-            ),
-
-          );
-        },
-      ),
-    ),
-  )
+                ),
+              );
+            },
+          ),
+        ),
+      )
       .toList();
 
   Widget _buildDropdown() {
     final overlayColor = MaterialStateProperty.resolveWith<Color?>((states) =>
-    states.contains(MaterialState.focused) ? Colors.transparent : null);
+        states.contains(MaterialState.focused) ? Colors.transparent : null);
     final iconStyleData = widget.icon != null
         ? IconStyleData(icon: widget.icon!)
         : const IconStyleData();
@@ -308,67 +307,66 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
       selectedItemBuilder: (context) => widget.options
           .map(
             (item) => Align(
-          alignment: AlignmentDirectional.centerStart,
-          child: Text(
-            isMultiSelect
-                ? currentValues
-                .where((v) => optionLabels.containsKey(v))
-                .map((v) => optionLabels[v])
-                .join(', ')
-                : optionLabels[item]!,
-            style: widget.textStyle,
-            maxLines: 1,
-          ),
-        ),
-
-      )
+              alignment: AlignmentDirectional.centerStart,
+              child: Text(
+                isMultiSelect
+                    ? currentValues
+                        .where((v) => optionLabels.containsKey(v))
+                        .map((v) => optionLabels[v])
+                        .join(', ')
+                    : optionLabels[item]!,
+                style: widget.textStyle,
+                maxLines: 1,
+              ),
+            ),
+          )
           .toList(),
       dropdownSearchData: widget.isSearchable
           ? DropdownSearchData<T>(
-        searchController: _textEditingController,
-        searchInnerWidgetHeight: 50,
-        searchInnerWidget: Container(
-          height: 50,
-          padding: const EdgeInsets.only(
-            top: 8,
-            bottom: 4,
-            right: 8,
-            left: 8,
-          ),
-          child: TextFormField(
-            expands: true,
-            maxLines: null,
-            controller: _textEditingController,
-            cursorColor: widget.searchCursorColor,
-            style: widget.searchTextStyle,
-            decoration: InputDecoration(
-              isDense: true,
-              contentPadding: const EdgeInsets.symmetric(
-                horizontal: 10,
-                vertical: 8,
+              searchController: _textEditingController,
+              searchInnerWidgetHeight: 50,
+              searchInnerWidget: Container(
+                height: 50,
+                padding: const EdgeInsets.only(
+                  top: 8,
+                  bottom: 4,
+                  right: 8,
+                  left: 8,
+                ),
+                child: TextFormField(
+                  expands: true,
+                  maxLines: null,
+                  controller: _textEditingController,
+                  cursorColor: widget.searchCursorColor,
+                  style: widget.searchTextStyle,
+                  decoration: InputDecoration(
+                    isDense: true,
+                    contentPadding: const EdgeInsets.symmetric(
+                      horizontal: 10,
+                      vertical: 8,
+                    ),
+                    hintText: widget.searchHintText,
+                    hintStyle: widget.searchHintTextStyle,
+                    border: OutlineInputBorder(
+                      borderRadius: BorderRadius.circular(8),
+                    ),
+                  ),
+                ),
               ),
-              hintText: widget.searchHintText,
-              hintStyle: widget.searchHintTextStyle,
-              border: OutlineInputBorder(
-                borderRadius: BorderRadius.circular(8),
-              ),
-            ),
-          ),
-        ),
-        searchMatchFn: (item, searchValue) {
-          return (optionLabels[item.value] ?? '')
-              .toLowerCase()
-              .contains(searchValue.toLowerCase());
-        },
-      )
+              searchMatchFn: (item, searchValue) {
+                return (optionLabels[item.value] ?? '')
+                    .toLowerCase()
+                    .contains(searchValue.toLowerCase());
+              },
+            )
           : null,
       // This is to clear the search value when you close the menu
       onMenuStateChange: widget.isSearchable
           ? (isOpen) {
-        if (!isOpen) {
-          _textEditingController.clear();
-        }
-      }
+              if (!isOpen) {
+                _textEditingController.clear();
+              }
+            }
           : null,
     );
   }
