+++ pubspec.yaml
@@ -2,26 +2,6 @@ name: flutterflow_ui
 description: Flutter package that makes it easy to use UI widgets and code generated by FlutterFlow in your Flutter projects.
 homepage: https://docs.flutterflow.io
 repository: https://github.com/FlutterFlow/flutterflow-ui
-
-# google_mobile_ads, flutter_braintree, and chat have been disabled.
-# flutter_html, sign_in_with_apple, and assets_audio_player have been added manually.
-
-
-
-# The following line prevents the package from being accidentally published to
-# pub.dev using `pub publish`. This is preferred for private packages.
-# publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 0.2.0+1
 
 environment:
@@ -65,29 +45,11 @@ dependencies:
   percent_indicator: 4.2.2
   rive: 0.12.2
   smooth_page_indicator:
+  shared_preferences: 2.2.2
   swipeable_card_stack: 0.0.5
   timeago: 3.2.2
   table_calendar: 3.0.9
   provider: 6.0.4
 
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-
-  # To add assets to your application, add an assets section, like this:
-
-
-
